refactor ♻️: Rename AI module and update import paths
This commit renames the 'git_acp.ai.generation' module to 'git_acp.commit.generation' and updates all references in '__init__.py'. Additionally, it refactors the PR builder to ensure consistent behavior. The changes are aimed at improving code organization and clarity while maintaining backward compatibility with existing functionality.
